+++ contracts/governance/GovernorTokenCreate.sol
@@ -82,7 +82,7 @@ contract GovernorTokenCreate is GovernorCountingSimpleInternal {
     function mintToken(
         uint256 proposalId,
         uint256 amount
-    ) external returns (int256) {
+    ) external returns (int64) {
         TokenCreateData storage tokenCreateData = _proposalData[proposalId];
 
         require(
@@ -95,7 +95,7 @@ contract GovernorTokenCreate is GovernorCountingSimpleInternal {
             "GovernorTokenCreate: Only treasurer can mint tokens."
         );
 
-        (int256 responseCode, int256 newTotalSupply) = super.mintToken(
+        (int256 responseCode, int64 newTotalSupply) = super.mintToken(
             tokenService,
             tokenCreateData.newTokenAddress,
             amount
@@ -111,7 +111,7 @@ contract GovernorTokenCreate is GovernorCountingSimpleInternal {
     function burnToken(
         uint256 proposalId,
         uint256 amount
-    ) external returns (int256) {
+    ) external returns (int64) {
         TokenCreateData storage tokenCreateData = _proposalData[proposalId];
         require(
             tokenCreateData.newTokenAddress != address(0x0),
@@ -123,7 +123,7 @@ contract GovernorTokenCreate is GovernorCountingSimpleInternal {
             "GovernorTokenCreate: Only treasurer can burn tokens."
         );
 
-        (int256 responseCode, int256 newTotalSupply) = super.burnToken(
+        (int256 responseCode, int64 newTotalSupply) = super.burnToken(
             tokenService,
             tokenCreateData.newTokenAddress,
             amount
@@ -160,7 +160,7 @@ contract GovernorTokenCreate is GovernorCountingSimpleInternal {
             address(this)
         );
         require(
-            contractBalance >= uint256(amount),
+            contractBalance >= amount,
             "GovernorTokenCreate: Contract doesn't have sufficient balance please take treasurer help to mint it."
         );
